Main map has been added.
The robot has been added to the Team8 Character Blueprint.

Left/Right Movement Works.
Jumping works.
Crouching Slows the character.
Attacks worked previously but the anim Blueprint needs to be modified so collision works again.

To be Changed/added:

The landscape needs to be modified to make sense for the map.
Textures needs to be added to the map.
Robot needs animations.
New camera system needs to be implemented.

Tem and Matt G completed the above tasks and merged projects and blueprints together.